+++ ReleaseNotes.md
@@ -3,6 +3,10 @@
 ## Features
 - Strikes can now be updated in added to RaidOrga+ via update Termin
 
+## Improvements
+- Add support for new two new RaidOrga+ roles Alacrity and Quickness
+- Rework role detection for update RaidOrga+ to support multiple roles
+
 ## Bugfixes
 - Fix issue with certain characters when parsing log locally